workers/requirements.txt
# AI Editor 2.0 Python Workers
# Redis Queue for background job processing

# Core dependencies
redis==5.0.1
rq==1.16.0
rq-scheduler==0.13.1

# Airtable API
pyairtable==2.3.3

# AI APIs
anthropic==0.39.0
google-generativeai==0.8.3
openai==1.55.0

# HTTP requests
requests==2.32.3
httpx==0.27.2

# Image processing
Pillow==10.4.0
cloudinary==1.41.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# HTML processing
beautifulsoup4==4.12.3
html2text==2024.2.26
